+++ src/Feature.ts
@@ -1,25 +1,26 @@
 import { map } from 'lodash';
+
 import { FeatureBuilder } from './FeatureBuilder';
 import { getGlobalTestMethods } from './lib/getGlobalTestMethods';
-import { IGherkinParserConfig, parseFeature } from './parseFeature';
+import { IGherkinParserConfig, parseFeature } from './lib/parseFeature';
 import {
-  IGherkinAst,
-  IGherkinAstEntity,
-  IGherkinAstStep,
-  IGherkinBackground,
-  IGherkinFeatureTest,
-  IGherkinMethods,
-  IGherkinOperationStore,
-  IGherkinScenario,
-  IGherkinStep,
-  IHookFn,
-  IMatch,
+  IGherkinAst, IGherkinAstEntity, IGherkinAstStep, IGherkinBackground, IGherkinFeatureTest, IGherkinMethods,
+  IGherkinOperationStore, IGherkinScenario, IGherkinStep, IGherkinTestSupportFlags, IHookFn, IMatch,
 } from './types';
 
 export type IConfigureFn = (t: IGherkinMethods) => void;
 
-export type IDescribe = (title: string, fn: () => any) => void;
-export type ITest = (name: string, fn: () => Promise<any>|any, timeout?: number) => void;
+export interface IDescribe {
+  (name: string, fn: () => Promise<any>|any): void;
+  skip: IDescribe;
+  only: IDescribe;
+}
+
+export interface ITest {
+  (name: string, fn: () => Promise<any>|any, timeout?: number): void;
+  skip: ITest;
+  only: ITest;
+}
 
 export interface ITestMethods {
   describe: IDescribe;
@@ -33,6 +34,9 @@ export interface IGherkinTestParams extends IGherkinParserConfig {
   defaultTimeout?: number;
 }
 
+/**
+ * Create a gherkin test based on a feature file
+ */
 export function Feature (config: IGherkinTestParams | IGherkinTestParams['feature'], configure: IConfigureFn) {
   const { feature, methods = getGlobalTestMethods(), defaultTimeout } = <IGherkinTestParams> (
     typeof config === 'string'
@@ -47,9 +51,6 @@ export function Feature (config: IGherkinTestParams | IGherkinTestParams['featur
   return { ast, featureBuilder };
 }
 
-/**
- * Create a gherkin test based on a feature file
- */
 export const GherkinTest = Feature;
 
 export type IOnConfigured = (callback: () => void) => void;
@@ -125,7 +126,9 @@ function describeFeature ({ featureBuilder, ast, configure, methods, defaultTime
 }
 
 /** Instruments step functions in the test framework */
-function describeGherkinOperations ({ state, steps, methods, defaultTimeout }: {
+function describeGherkinOperations ({
+  state, steps, methods, defaultTimeout,
+}: {
   steps: IGherkinOperationStore,
   state: PromiseLike<any>;
   methods: ITestMethods;
@@ -137,9 +140,13 @@ function describeGherkinOperations ({ state, steps, methods, defaultTimeout }: {
     const title = formatTitle({ name: step.name });
     const timeout = step.timeout || defaultTimeout;
 
-    methods.test(title, async () => {
+    const execute = async () => {
       reducedState = await executeStep(step, reducedState);
-    }, timeout);
+
+      return reducedState;
+    };
+
+    methods.test(title, execute, timeout);
   });
 }
 
@@ -169,10 +176,14 @@ function describeScenario ({
   defaultTimeout?: number;
 }) {
   const title = formatTitle(scenario.gherkin);
+  const { skip, only } = scenario;
+  const describeMethod = narrowTestMethod(methods.describe, { skip, only });
 
-  methods.describe(title, () => {
-    afterEachHooks.forEach(({ fn, timeout = defaultTimeout }) => { methods.afterAll(fn, timeout); });
-    beforeEachHooks.forEach(({ fn, timeout = defaultTimeout }) => { methods.beforeAll(fn, timeout); });
+  describeMethod(title, () => {
+    if (!skip) {
+      afterEachHooks.forEach(({ fn, timeout = defaultTimeout }) => { methods.afterAll(fn, timeout); });
+      beforeEachHooks.forEach(({ fn, timeout = defaultTimeout }) => { methods.beforeAll(fn, timeout); });
+    }
 
     const scenarioSteps = new Map([
       ...background && background.Given
@@ -185,6 +196,13 @@ function describeScenario ({
 
     describeGherkinOperations({ steps: scenarioSteps, state: initialState, methods });
   });
+}
+
+function narrowTestMethod (method: ITest, { skip, only }: IGherkinTestSupportFlags) {
+  if (only) { return method.only; }
+  if (skip) { return method.skip; }
+
+  return method;
 }
 
 function augmentBackgroundStepNames (steps: Map<IMatch, IGherkinStep<IGherkinAstStep>>) {